+++ packages/client/src/Call.ts
@@ -372,6 +372,9 @@ export class Call {
       this.registerEffects();
       this.registerReconnectHandlers();
 
+      // Set up the device managers again. Although this is already done
+      // in the DeviceManager's constructor, they'll need to be re-set up
+      // in the cases where a call instance is recycled (join -> leave -> join).
       this.camera.setup();
       this.microphone.setup();
       this.screenShare.setup();
@@ -638,6 +641,8 @@ export class Call {
       this.statsReporter?.stop();
       this.statsReporter = undefined;
 
+      const leaveReason = message ?? reason ?? 'user is leaving the call';
+      this.tracer.trace('call.leaveReason', leaveReason);
       this.sfuStatsReporter?.flush();
       this.sfuStatsReporter?.stop();
       this.sfuStatsReporter = undefined;
@@ -648,9 +653,7 @@ export class Call {
       this.publisher?.dispose();
       this.publisher = undefined;
 
-      await this.sfuClient?.leaveAndClose(
-        message ?? reason ?? 'user is leaving the call',
-      );
+      await this.sfuClient?.leaveAndClose(leaveReason);
       this.sfuClient = undefined;
       this.dynascaleManager.setSfuClient(undefined);
       await this.dynascaleManager.dispose();
@@ -857,6 +860,7 @@ export class Call {
    * Unless you are implementing a custom "ringing" flow, you should not use this method.
    */
   accept = async () => {
+    this.tracer.trace('call.accept', '');
     return this.streamClient.post<AcceptCallResponse>(
       `${this.streamClientBasePath}/accept`,
     );
@@ -874,6 +878,7 @@ export class Call {
   reject = async (
     reason: RejectReason = 'decline',
   ): Promise<RejectCallResponse> => {
+    this.tracer.trace('call.reject', reason);
     return this.streamClient.post<RejectCallResponse, RejectCallRequest>(
       `${this.streamClientBasePath}/reject`,
       { reason: reason },
@@ -1850,12 +1855,6 @@ export class Call {
         await this.publisher.publish(audioTrack, screenShareAudio, options);
         trackTypes.push(screenShareAudio);
       }
-    }
-
-    if (track.kind === 'video') {
-      // schedules calibration report - the SFU will use the performance stats
-      // to adjust the quality thresholds as early as possible
-      this.sfuStatsReporter?.scheduleOne(3000);
     }
 
     await this.updateLocalStreamState(mediaStream, ...trackTypes);