+++ packages/client/src/coordinator/connection/types.ts
@@ -108,17 +108,35 @@ export type ConnectionRecoveredEvent = {
   type: 'connection.recovered';
 };
 
+export type MicCaptureReportEvent = {
+  type: 'mic.capture_report';
+  call_cid: string;
+  /**
+   * Whether the mic is capturing audio.
+   */
+  capturesAudio: boolean;
+  /**
+   * The device ID of the mic.
+   */
+  deviceId?: string;
+  /**
+   * The human-readable label of the mic.
+   */
+  label?: string;
+};
+
 export type StreamVideoEvent = (
   | VideoEvent
   | NetworkChangedEvent
   | ConnectionChangedEvent
   | TransportChangedEvent
   | ConnectionRecoveredEvent
+  | MicCaptureReportEvent
 ) & { received_at?: string | Date };
 
 // TODO: we should use WSCallEvent here but that needs fixing
 export type StreamCallEvent = Extract<StreamVideoEvent, { call_cid: string }>;
-export type EventTypes = 'all' | VideoEvent['type'];
+export type EventTypes = 'all' | StreamVideoEvent['type'];
 
 export type AllClientEventTypes = 'all' | StreamVideoEvent['type'];
 export type AllClientEvents = {
@@ -129,7 +147,7 @@ export type ClientEventListener<E extends keyof AllClientEvents> = (
 ) => void;
 
 export type AllClientCallEvents = {
-  [K in EventTypes]: Extract<VideoEvent, { type: K }>;
+  [K in EventTypes]: Extract<StreamVideoEvent, { type: K }>;
 };
 
 export type AllCallEvents = AllClientCallEvents & AllSfuEvents;